Component Code Refracted
