+++ frontend/src/common/form.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import Joi from "joi-browser";
 import Select from "./select";
 import Input from "./input";
