+++ frontend/src/forms/index.jsx
@@ -1,0 +1,2 @@
+export { default as SigninForm } from "./SigninForm/SigninForm";
+export { default as SignupForm } from "./SignupForm/SignupForm";
